refactor: extract IssueExportPresentationController.swift from IssueExportPresenters.swift
Splits the `IssueExportPresentationController` @MainActor class (~48
lines) out of IssueExportPresenters.swift so that file holds only the
`IssueExportStatusPresenter` enum.

Byte-identical move of lines 21-69. Both files import Foundation.

IssueExportPresenters.swift: 69 → 19 lines. Tests: 667.
